requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==2.0.1
ollama==0.3.1
openai==1.41.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
regex==2024.7.24
requests==2.32.3
scipy==1.14.0
sniffio==1.3.1
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
